==> 2. Data cleaning with Python.ipynb <==
""" Data wrangling with Python for the 5 data tables that are part of the database """

import pandas as pd

# importing the five CSVs

missing_values = ["-", "--", "n/a", "na", "NA", "N/A", "?", "None", "null", "NULL", ""] 

buyers = pd.read_csv(r"C:\Users\Desktop\Data Sets\(2)\buyers.csv", na_values=missing_values)
products = pd.read_csv(r"C:\Users\\Desktop\Data Sets\(2)\products.csv", na_values=missing_values)
reviews = pd.read_csv(r"C:\Users\\Desktop\Data Sets\(2)\reviews.csv", na_values=missing_values)
sellers = pd.read_csv(r"C:\Users\Desktop\Data Sets\(2)\sellers.csv", na_values=missing_values)
transactions = pd.read_csv(r"C:\Users\Desktop\Data Sets\(2)\transactions.csv", na_values=missing_values)

# Cleaning the BUYERS table

buyers.head()

buyers.tail()

buyers.shape

buyers.info()

buyers.dtypes

buyers["birthdate"] = pd.to_datetime(buyers["birthdate"])
buyers["gender"] = buyers["gender"].astype("category")
buyers["country"] = buyers["country"].astype("category")
buyers["customer_type"] = buyers["customer_type"].astype("category")
buyers["buyer_id"] = buyers["buyer_id"].astype("object")

def clean_buyers(df):
    df["birthdate"] = pd.to_datetime(df["birthdate"])
    df["gender"] = df["gender"].astype("category")
    df["country"] = df["country"].astype("category")
    df["customer_type"] = df["customer_type"].astype("category")
    df["buyer_id"] = df["buyer_id"].astype("object")


for x in buyers.columns:
    value_count_buyers = buyers[x].value_counts()
    duplicated_buyers = value_count_buyers[value_count_buyers > 1]
    print(duplicated_buyers)

buyers[buyers.duplicated(subset=["name"],keep=False)].sort_values(by=["name"])

# you can have an overview of every duplicate (inc the "original") by using keep=False
# you can find duplicates by subsets

buyers["birthdate"].sort_values() # Min: 1944 // Max: 2007

buyers["name"] = buyers["name"].str.title().str.strip()
buyers["country"] = buyers["country"].str.title().str.strip()
buyers["customer_type"] = buyers["customer_type"].str.lower().str.strip()

for col in buyers.columns:
    unique_vals = buyers[col].unique()
    print(unique_vals)

buyers["birthdate"] = pd.to_datetime(buyers["birthdate"], errors="coerce")
buyers[buyers["birthdate"].isnull()]

# If 'coerce', then invalid parsing will be set as NaT
# No errors were recorded when changing types

buyers["birthdate"].info()

for col in buyers.columns:
    x = buyers[col].nunique()
    print(col, x)

# Cleaning the PRODUCTS table

products.head()

products = products.rename(columns={"book ID": "book_id"})

products.tail()

products.shape

products.info()

products.describe()

products["category"] = products["category"].str.lower().str.strip()
products["rating"] = products["rating"].str.lower().str.strip()

products["category"] = products["category"].astype("category")
products["rating"] = products["rating"].astype("category")

products.dtypes

for x in products.columns:
    value_count_products = products[x].value_counts()
    duplicated_products = value_count_products[value_count_products > 1]
    print(duplicated_products)

print(products["category"].value_counts(dropna=False))

print(products["rating"].value_counts(dropna=False))

products[products["seller_id"] == 39]

products["price"].sort_values()

products.head()

# Cleaning the REVIEWS table

reviews.head()

reviews.tail()

reviews.shape

reviews.info()

# null values in review_text that need to be cleaned

reviews.dtypes

reviews["review_date"] = pd.to_datetime(reviews["review_date"], errors="coerce")
reviews["review_id"] = reviews["review_id"].astype("object")

# some values cannot be converted to integer, let's take care of these values first

reviews["rating"].unique()
reviews["rating"] = reviews["rating"].str.replace("Three", "3")
reviews["rating"] = reviews["rating"].str.replace("Bad", "1")
reviews["rating"].unique()

reviews["rating"] = reviews["rating"].astype(int)

reviews["review_date"].info()

for x in reviews.columns:
    print(x, reviews[x].isnull().sum())

reviews[reviews["review_text"].isnull()]
reviews["review_text"] = reviews["review_text"].fillna("No review submitted")

for x in reviews.columns:
    value_count_reviews = reviews[x].value_counts()
    duplicated_reviews = value_count_reviews[value_count_reviews > 1]
    print(duplicated_reviews)

reviews["book_id"].nunique() # 13
reviews["author"].nunique() # Too many authors generated by ChatGPT 

# Cleaning the SELLERS table

sellers.head()

sellers.tail()

sellers.shape

sellers.info()

sellers.dtypes

sellers["author_name"] = sellers["author_name"].str.title().str.strip()
sellers["country"] = sellers["country"].str.title().str.strip()
sellers["seller_type"] = sellers["seller_type"].str.lower().str.strip()

sellers["seller_id"] = sellers["seller_id"].astype("object")
sellers["birthdate"] = pd.to_datetime(sellers["birthdate"], errors="coerce")
sellers["country"] = sellers["country"].astype("category")
sellers["seller_type"] = sellers["seller_type"].astype("category")

sellers["birthdate"].info()

for x in sellers.columns:
    value_count_sellers = sellers[x].value_counts()
    duplicated_sellers = value_count_sellers[value_count_sellers > 1]
    print(duplicated_sellers)

sellers["country"].unique()

sellers["seller_type"].unique()

# Cleaning the TRANSACTIONS table

transactions.head()

transactions.tail()

transactions.shape

transactions.info()

# price_paid has null values !

transactions.describe()

# price_paid has negative values (credit note ?)

transactions.dtypes

transactions["transaction_id"] = transactions["transaction_id"].astype("object")
transactions["buyer_id"] = transactions["buyer_id"].astype("object")
transactions["date"] = pd.to_datetime(transactions["date"], errors="coerce")

transactions["date"].info()

for x in transactions.columns:
    print(x, transactions[x].nunique())

for x in transactions.columns:
    print(x, transactions[x].isnull().sum())

# 103 records in the column price_parid are null values

transactions.head()

transactions = transactions[transactions["price_paid"] > 0]
transactions["price_paid"].unique()

# deleting all the irrelevant transactions that aren't related to the list of my books

transactions = transactions.merge(products, how="inner", left_on="book_id", right_on="book_id")

transactions["book_id"].unique()

transactions.shape

transactions.info()

transactions["price_paid"] = transactions["price"]

list = ["name", "price", "category", "summary", "rating", "seller_id"]

for x in list:
    transactions = transactions.drop(x, axis=1)

transactions = transactions.rename(columns={"price_paid":"price"})

transactions.insert(4, "price_paid", transactions["price"] - (transactions["price"] * transactions["discount_rate"]))
transactions["price_paid"] = transactions["price_paid"].round(2)

transactions.head()

transactions.info()

is_unique_buyer_id = buyers["buyer_id"].is_unique
print("buyer_id :", is_unique_buyer_id) 
is_unique_book_id = products["book_id"].is_unique
print("book_id :", is_unique_book_id)
is_unique_review_id = reviews["review_id"].is_unique
print("review_id :", is_unique_review_id)
is_unique_seller_id = sellers["seller_id"].is_unique
print("seller_id :", is_unique_seller_id)
is_unique_transaction_id = transactions["transaction_id"].is_unique
print("transaction_id :", is_unique_transaction_id)


missing_values_transactions_book_id_fkey = transactions[~transactions["book_id"].isin(products["book_id"])]
print(f"FK book_id missing in transactions: {len(missing_values_transactions_book_id_fkey)}")
missing_values_transactions_buyer_id_fkey = transactions[~transactions["buyer_id"].isin(buyers["buyer_id"])]
print(f"FK buyer_id missing in transactions: {len(missing_values_transactions_buyer_id_fkey)}")
missin_values_reviews_book_id_fkey2 = reviews[~reviews["book_id"].isin(products["book_id"])]
print(f"FK book_id missing in reviews: {len(missin_values_reviews_book_id_fkey2)}")


# both = match in both tables 
# left_only = exists in transactions but not in products (foreign key absent)
# right_only = exists in products but not in transactions

merged = transactions.merge(products, how="left", on="book_id", indicator=True)
merged[merged["_merge"] == "left_only"]

merged = transactions.merge(buyers, how="left", on="buyer_id", indicator=True)
merged[merged["_merge"] == "left_only"]

merged = reviews.merge(products, how="left", on="book_id", indicator=True)
merged[merged["_merge"] == "left_only"]

buyers.to_csv("results_python/buyers.csv", index=False)
products.to_csv("results_python/products.csv", index=False)
reviews.to_csv("results_python/reviews.csv", index=False)
sellers.to_csv("results_python/sellers.csv", index=False)
transactions.to_csv("results_python/transactions.csv", index=False)
